More work has been done fixing memory leaks. Now, a few remain, but these are apparent only when writing lots of groups/leaves (typically 3 MB/ 200 tables). However, for reading it seems that there are no leaks!.
Also, a lazy reading of attributes is now the default. This will speed-up
the reading of large trees a factor of 25% more or less.

Writing rows in tables has been also optimized (a 15% of improvement).

Reading is finally using the optimization in Pyrex. With that, the same
performance than 0.7 is achieved. The fix in memory leaks seems to have
reduced the performance in reading. However, with this, the performance is
more or less the same.

Code has been beautified (hdf5Extension.pyx and Table.py)


git-svn-id: http://www.pytables.org/svn/pytables/trunk@168 1b98710c-d8ec-0310-ae81-f5f2bcd8cb94