The Table.read method has been reworked, and some special cases has been optimized, most specially when reading a complete table in-memory with a step=1. In that case a factor three of improvement has been achieved in comparison with the general step=n case. In the special case of step=1, 160 MB/s (!) in reading tables can be achieved now.
It also has been implemented a Pyrex version for the case that only one
column is required and step=1, but in that case the improvement can only be
seen when the column to be read is of type String (a 10%), but when the
columns are integers or floats, the performance can be twice as bad (!), so
this optimization has been commented out in the final code.

Found as well a couple of bugs in the included HDF5_HL library
(H5TBread_fields_names and H5TBread_fields_index). I'm reporting these bugs
to NCSA.


git-svn-id: http://www.pytables.org/svn/pytables/trunk@174 1b98710c-d8ec-0310-ae81-f5f2bcd8cb94