Added changes to numexpr by Francesc Altet and integrated into PyTables.
The changes to numexpr are summarised under `tables/numexpr/README.txt`.

For integration, the `tables.numexpr` package has been added to the manifest
file, numexpr's license is included in the `LICENSES` directory, and `setup.py`
includes the new subpackage and builds the `interpreter` extension.  Numexpr’s
own `setup.py` is not included since it is no longer necessary.


git-svn-id: http://www.pytables.org/svn/pytables/branches/professional@1606 1b98710c-d8ec-0310-ae81-f5f2bcd8cb94