Made the next optimizations:
- When a table has not nested columns, it will return a plain
numarray.RecArray instead of a tables.nestedrecords.NestedRecArray.
I've done this because the handling of the later objects is quite
faster than the former.

- A new Table._v_rbuffer property has been defined as a buffer reding
cache. When a reading method of Table has to return a RecArray, it
calls this property and, if the resulting number of rows is less than
this buffer, then it gets a subrange of it. This operation is much
faster (up to 10x) than creating a fresh recarray.

These optimizations are specially important in benchmarks when you
have to show PyTables Pro lookup capabilities in the sub-millisecond
range :-)



git-svn-id: http://www.pytables.org/svn/pytables/branches/professional@1653 1b98710c-d8ec-0310-ae81-f5f2bcd8cb94