Added initial version of Table.where() using numexpr.
The method is called Table.where2XXX(), and it is completely isolated from the
old where().  It is not complete, and it is there just for initial benchmarking
of the new query mechanism using numexpr.  It has been tested by tailoring the
calls to where() i `test_tables.py`.

This experimental code is marked with XXX all over the place.

Also, some importing problems with numexpr have been fixed by making the
imports absolute (need to reduce those absolute imports to not differ more than
necessary from mainstream numexpr).


git-svn-id: http://www.pytables.org/svn/pytables/branches/professional@1658 1b98710c-d8ec-0310-ae81-f5f2bcd8cb94