Added a cache of splitted conditions to `Table`.
It is completely handled by `Index.split_index_condXXX()`.  Splitted conditions
are not as static as one may expect, since:

1. The set of columns with *usable* indexes may vary from time to time (e.g.
   because they are dirty).
2. The values of variables which act as limits may change from invocation to
   invocation.

That's why the key of the cache is not only the condition itself, but also the
set of usable indexes, and the reason why the values of limit variables are not
stored, only their names.


git-svn-id: http://www.pytables.org/svn/pytables/branches/professional@1688 1b98710c-d8ec-0310-ae81-f5f2bcd8cb94