Added partial string support (only comparisons) to numexpr.
String arrays are not copied even if they are not contiguous.  (In fact, under
numarray they are because of shortcomings in `NA_InputArray()`.)  The Python
side of the implementation is similar to that of long integers and other types.
The C side includes new opcodes, computing item size for constants, inputs,
temporaries and return value.  There still are some rough corners (i.e. no
string temporaries or returns are allowed).  Check for "XXX" comments.


git-svn-id: http://www.pytables.org/svn/pytables/branches/professional@1714 1b98710c-d8ec-0310-ae81-f5f2bcd8cb94