- Improved searching time in indexed queries when the number of results is 0.
- Added some comments in Table.py about the problems that people is
reporting lately to the PyTables list about strange segfaults, double
frees, etc...  The problems seem to appear when appending to a table,
one of its columns is automatically (the default) indexed. The problem
is not solved yet, but I'm starting to get an idea of what's going on.
I hope to solve definitely the problem anytime soon.



git-svn-id: http://www.pytables.org/svn/pytables/branches/professional@1728 1b98710c-d8ec-0310-ae81-f5f2bcd8cb94