- Solved the bug exposed by test_indexvalues.py. It was a problem with the key in the LRU cache for sorted arrays (in the context of indexation, of course).
- A bit of more optimization in the LRU cache code in utilsExtension.
Now, it is very fine tuned for the PyTables need (and code there is no
longer useful as a general purpose LRU cache, so be careful!).

- The SORTED_CACHE_SIZE and BOUNDS_CACHE_SIZE have been increased from
256 to 1024 nodes. This will let the indexation benchmarks to fit
comfortably and runs smoothly. A problem here is that, when the data
of a search doesn't fit well in the LRU cache, the continuous feeding
of the cache with array objects (that will be preempted anyway before
using a single one!) does take a lot of CPU (for nothing). The problem
is very serious, and profiling says that the main bottleneck is the
array.copy() method (indeed, the array has to be copied before being
included in the cache). I've done some timings on whether numpy can
somewhat alleviate the problem, and it seems that copies in numpy are
around 20x times faster than numarray (arange(1000).copy() takes 52.3
us in numarray and a mere 2.6 us in numpy). So using numpy in the LRU
cache scenarios can be a *big* win. It will be also an enormeous win
in the context of searches that does return results (the creation of a
recarray buffer in numarray takes 170 us, while numpy takes just 5.8
us, which represents a 30x speed-up). Will this imply that we will
finally need to migrate towards using numpy in order to achieve decent
speeds in searches?.

- Now, all tests passes successfully, except some problems when using
numpy 1.0b1 that will hopefully be addressed soon.



git-svn-id: http://www.pytables.org/svn/pytables/branches/professional@1739 1b98710c-d8ec-0310-ae81-f5f2bcd8cb94