Added machinery to detect situations where the cache for numerical data is not effective in order to disable it. The algorithm is quite intelligent in that it will be re-enabled from time to time in order to see if it is useful again.
This has been implemented only for a special version of the LRU cache
that lives in indexesExtension.pyx for use in indexation scenarios.
With this, the LRU cache for numerical data is very efficient in all
the scenarios that it can. And if it is not, it will be disabled
automatically.



git-svn-id: http://www.pytables.org/svn/pytables/branches/professional@1740 1b98710c-d8ec-0310-ae81-f5f2bcd8cb94