Added a new cache behaviour for nodes: when NODE_CACHE_SIZE has a negative value, the _AliveNodes dictionary will create *strong* references (instead of weak refs) to node objects. That makes that all the touched nodes will go to the _AliveNodes dictionary and will live there forever. The user gets a PerformanceWarning when the number of alive nodes reaches -NODE_CACHE_SIZE size. When NODE_CACHE_SIZE is 0, it is like a negative value, but the user doesn't get warned.
So, for benchmarks, having a negative value for NODE_CACHE_SIZE (i.e.
all the nodes will be kept alive) is a better default than having the
LRU cache for nodes, because the former is about 5x times faster than
the later. However, for general use, it seems more sensible to keep a
positive value for NODE_CACHE_SIZE.



git-svn-id: http://www.pytables.org/svn/pytables/branches/professional@1776 1b98710c-d8ec-0310-ae81-f5f2bcd8cb94