Fixed splitted condition cache support.
The old cache only used the condition string as a key.  However, that an
already splitted condition can be used again also depends on the set of column
names, variable names, columns used, variable types and usable indexes.

Since the set of indexed columns does not vary quickly, and it affects the way
conditions are splitted, and checking it implies access to nodes, this set does
not appear in the key, but changing it simply invalidates the whole cache.
Individual conditions are not removed because it would be difficult to check in
which exact cached conditions which columns are being used.

Since the set of dirty indexed columns varies quickly, and a dirty index avoid
using an splitted condition, the presence of dirty indexes simply disables the
cache.  The new cache keeps a count of dirty indexes.


git-svn-id: http://www.pytables.org/svn/pytables/branches/professional@1798 1b98710c-d8ec-0310-ae81-f5f2bcd8cb94