Use compiled Numexpr functions in the condition cache and splitting code.
This should reduce the overhead of repeatedly doing the same query on a table,
since the residual expression does not go through `evaluate()` and it doesn't
need to be rebuilt into a string.  Now the condition splitting function returns
a compiled function and a parameter list so that the former can be reused
simply by calling it with the arguments in the proper order.  Moreover, knowing
exactly which variables take part in the expressions avoids having to read data
from columns which are not actually used.


git-svn-id: http://www.pytables.org/svn/pytables/branches/professional@1813 1b98710c-d8ec-0310-ae81-f5f2bcd8cb94