- Migrated EArray.py and VLArray.py modules. Also, test_earray.py has been migrated and all tests passes there. I haven't done anything to test_vlarray.py yet.
- Due to some different ways of computing an integer division of
negative numbers in Python and C, the way for computing the number has
been changed. In Python now it is used len(xrange(start,stop,step))
and in C, the responsible for doing the computation is a new function
in utils.c called get_len_of_range (copied and slightly changed from
Python itself). For more info about this issue, see:

http://www.python.org/doc/faq/programming.html#why-does-22-10-return-3

and

http://www.peterbe.com/Integer-division-in-programming-languages




git-svn-id: http://www.pytables.org/svn/pytables/branches/professional@1834 1b98710c-d8ec-0310-ae81-f5f2bcd8cb94