Implemented a new way of extending the dictionary of condition variables.
The new `Table._requiredExprVars()`, which replaces `_extendedWithDefVars()',
compiles the condition to get the names of the variables used in it.  This
allows to "pull" only the needed ones, check that they are properly defined,
make early checks on the kinds of columns and save checking for identifier-like
column names (if the variable is in the expression, it is a valid indentifier).


git-svn-id: http://www.pytables.org/svn/pytables/branches/professional@1858 1b98710c-d8ec-0310-ae81-f5f2bcd8cb94