Automagically get condition variables from the user's namespace.
See the docstring of `Table._requiredExprVars()` for the restrictions on its
usage that this new mechanism imposes on API callables implementing it.  The
documentation should make clear that this is intended for interactive usage,
and that it can be disabled by providing a (possibly empty) mapping for the
condition variables (instead of the new default `None`).


git-svn-id: http://www.pytables.org/svn/pytables/branches/professional@1860 1b98710c-d8ec-0310-ae81-f5f2bcd8cb94