Removed the `Table.removeIndex()` method.
This was redundant since there was already a `Column.removeIndex()`.  Moreover,
there was a `Column.createIndex()` method but not a `Table.createIndex()` one.
To top it off, it is clearly a column operation.

By the way, internal update of `Table.colindexed` and `Table.indexed`
attributes has been factored into `Table._setColumnIndexing()` (so that
`Column` instances tamper a little less with `Table` attributes).


git-svn-id: http://www.pytables.org/svn/pytables/branches/professional@1906 1b98710c-d8ec-0310-ae81-f5f2bcd8cb94