Refactorization of *Array._loadEnum done. Now, it has been removed from hdf5Extension.Array and hdf5Extension.VLArray and unified and put in hdf5Extension.Leaf. Perhaps I'd have done the same with TableExtension.Table._loadEnum, but I couldn't because TableExtension.Table doesn't inherits yet from hdf5Extension.Leaf :-(
As a consequence of refactorization, the byteorder of enumerated types
is computed directly in _loadEnum, and incorporated into the .dtype
natively (before, there was a separation between Leaf.byteorder and
Leaf.dtype.byteorder for enumerated types). I think that the new
situation is much more consistent.

All the tests pass now again (except those in test_queries ;-)



git-svn-id: http://www.pytables.org/svn/pytables/branches/professional@1936 1b98710c-d8ec-0310-ae81-f5f2bcd8cb94