Forbid indexing and querying `UInt64` columns.
(This commit completes r1939, which I couldn't avoid committing by mistake.)

The reason is that Numexpr can not safely convert `uint64` arrays into `int64`
(aka `long`) ones.  Since this may be supported in the future, a
`NotImplementedError` is raised.  Tests have been changed accordingly.

One may be tempted to enable indexing `UInt64` columns, since expressions like
`cUInt64 > x` still work if the column is indexed.  However, when the index can
not be used, the column must be passed to Numexpr, so the problem still raises,
and it is safer and less surprise-prone to avoid `UInt64` indexation
altogether.


git-svn-id: http://www.pytables.org/svn/pytables/branches/professional@1941 1b98710c-d8ec-0310-ae81-f5f2bcd8cb94