More refactoring in chunksize/maxTuples computation so that now, all the computational work is centralized in:
Array._calcChunkshape(self, expectedrows) and
Array._calcMaxTuples(self, expectedrows)

for all Array, CArray and EArray.

Note that there is not a real need to define _calcChunkshape() method
in Array because it is based on a non-chunked dataset. It is there
just as a way to share the definition between EArray and CArray. In
the future, this should be moved into CArray (and make EArray inherit
from CArray).

Along the factorization, I've managed to define three new properties
that are common to all Array, CArray and EArray. These are:

nrows -- The length of the first dimension of the array.
rowsize -- The size (in bytes) of the row in dimensions orthogonal to maindim.
maindim -- The dimension on which iterators do work. (new)

This has been possible to the introduction of a new and useful
concept: the main dimension. Such a main dimension refers to the
dimension along which the array iterators do work. Its introduction
does simplifies many things inside the code as well.



git-svn-id: http://www.pytables.org/svn/pytables/branches/professional@2031 1b98710c-d8ec-0310-ae81-f5f2bcd8cb94