Merged the 'pro-newatom' branch into the 'professional' branch.  Also, test_numarray.py has been ported to use the new atom system.  All the tests pass now.
Note: because of complexity reasons, the usersguide.xml file has not
been merged in this changeset. This will be addressed later on.

Note 2: because of limitations of svn working over the https protocol,
it seems that tables/atom.py could not be added (copied) directly in
this changeset. I'll remove it now, and try to add it afterwards.



git-svn-id: http://www.pytables.org/svn/pytables/branches/professional@2126 1b98710c-d8ec-0310-ae81-f5f2bcd8cb94