Converted `atom_from_*()` and `col_from_*()` into factory methods.
This allows the user to have nearly all the functionality in the `atom.py`
module (and the coresponding part of `IsDescription.py`) just by importing the
`Atom` class (or the `Col` class).  This also reduces the number of symbols exported by PyTables.

Now the following code:

{{{
>>> atom = atom_from_kind('int', 8)
>>> col = col_from_kind('int', 8, pos=4)
}}}

is written like this:

{{{
>>> atom = Atom.from_kind('int', 8)
>>> col = Col.from_kind('int', 8, pos=4)
}}}


git-svn-id: http://www.pytables.org/svn/pytables/branches/professional@2208 1b98710c-d8ec-0310-ae81-f5f2bcd8cb94