Merged `auto` and `reindex` on `IndexProps` into `auto`.
The new indexing code made the difference quite subtle, and maybe confusing to
the user.

By the way, `Index.nelements` i `nelementsLR` are not set to 0 when the index
becomes dirty.  This was purely aesthetical, but caused some inconsistencies on
values when reopening files.  Moreover, better not to lie about the size of an
index!


git-svn-id: http://www.pytables.org/svn/pytables/branches/professional@2264 1b98710c-d8ec-0310-ae81-f5f2bcd8cb94