Fixed assumption that `_i_table` only exists when `table` is indexed.
This assumption is wrong when moving or removing a table, and may lead to stale
group indexes.  So the existence of the index group is checked instead of
whether the table is indexed.  The check is performed by a `try` statement to
avoid `File.__contains__()` and then `File._getNode()`, which are way slower.


git-svn-id: http://www.pytables.org/svn/pytables/branches/professional@2281 1b98710c-d8ec-0310-ae81-f5f2bcd8cb94