Solved some problems with the byteorder setting in some special types (namely, 'bool', 'time' & 'string'). With this, problems with time should be fixed, although carrying out some tests in that regard would be nice.
The code for detecting the byteorder of tables has been enhanced to
look into nested types as well.

Incidentally, mixed byteorders columns in the same table seems to be
supported now (only for opening). So, no error is raised now when the
situation is detected, but rather table.desc._v_byteorder (and hence,
table.byteorder) is set to 'mixed'. This 'feature' will be not
documented, as it is strange enough and I don't like the perpective of
people trusting on it.



git-svn-id: http://www.pytables.org/svn/pytables/branches/professional@2320 1b98710c-d8ec-0310-ae81-f5f2bcd8cb94