Added some magic so that `IntAtom()` and the like can create instances.
The instances created are actually of a concrete class.  This is achieved by
changing the class of the instance being initialised, which is clearly a hack,
and the creation of `Col` subclasses must now take this into account.

So, contrary to [2338], abstract classes are now exported again, which allows
to keep table description declarations more uniform when mixing columns of
known kinds with both known and unknown item sizes.


git-svn-id: http://www.pytables.org/svn/pytables/branches/professional@2345 1b98710c-d8ec-0310-ae81-f5f2bcd8cb94