Disable some caches when `lrucacheExtension` is not available.
This should help to take this module out of the Standard version.  The affected
objects are:

 * `Table.sparsecache` and `Table.limdatacache`, which may be `None`.
 * The `_DeadNodes` class in `File.py` may be a normal `LRUCache`.

Some Python wrapper methods have been added for `NumCache`, since it may be
used in `tableExtension` (as `Table.sparsecache`) while not being declared.
This may have an impact on speed, and its importance should be profiled.

Part of the `setup.py` file has been rewritten to use the `lrucacheExtension`
only when available.


git-svn-id: http://www.pytables.org/svn/pytables/branches/professional@2347 1b98710c-d8ec-0310-ae81-f5f2bcd8cb94