Removed type, dtype and itemsize as properties of *Array classes. Now, these attributes should be found in atom.type, atom.dtype and atom.itemsize respectively. All docs and tests should have been updated.
A lot of refactorization has been done in order to take advantage of
the uniformity that allows the atom attribute in all the *Array
classes. VLArray has been refactorized as well, but more can be done
(in particular, removing the _atomicdtype, _atomictype, _atomicshape
and _atomicsize can be replaced by scatom or atom properties). Well,
this will have to wait.
	    



git-svn-id: http://www.pytables.org/svn/pytables/branches/professional@2386 1b98710c-d8ec-0310-ae81-f5f2bcd8cb94