Implemented in ptrepack utility the possibility to regenerate old indexes coming from 1.x format to the new 2.x format.
Also, discovered a flaw in the inheritance of the filters properties
of filters of indexes. Before, as table.indexFilters will be always
set, this will always make the filters of the index to be the same
of the table, even if the table *doesn't* have any! This has been
corrected.

path.py:joinPath has received a minor enhancement: it accepts now
complete pathnames as `name` parameter.




git-svn-id: http://www.pytables.org/svn/pytables/branches/professional@2443 1b98710c-d8ec-0310-ae81-f5f2bcd8cb94