The Index.reorder_slices() method has been completely rewritten from scratch. Now, *two* complete slices are taken at a time (including the last row slice) so as to sort them.  Then, each new slice that is read is put at the end of this two-slice buffer, while the previous one is moved to the beginning of the buffer. This is in order to make the last row to participate in the whole index reordering (and hence, significantly reducing the index entropy). Also, the new algorithm seems to be better at reducing the entropy of the regular part (i.e. all except the last row) of the index.
The LastRow datasets in indexes are now CArrays instead of plain Array
objects. This allows them to be compressed.

The begin/end bounds cache in LastRow have been moved from the
beginning of the sorted dataset to end, so that actual data starts now
at the beginning of the dataset.

The computation of the slicesize has been refined to both adapt to the
new Index.reorder_slices() and to better adapt to the different values
of the memlevel argument of the Col.createIndex().




git-svn-id: http://www.pytables.org/svn/pytables/branches/professional@2682 1b98710c-d8ec-0310-ae81-f5f2bcd8cb94