Based on experiments done for a discussion in the hdf-forum list, a list of improvements have been added to the ObjectCache and NumCache list (lrucacheExtension):
- Now, the time that the cache will sense its efficiency before being
considered to be disabled is fixed by the new parameter
DISABLE_EVERY_CYCLES.  This allows for finer control over the
disable/enable cicly of the caches.

- The NumCache class has been highly simplified and the key lookup
method has changed to a dictionary lookup instead of a binary search
in a sorted list.  The new lookup method should be 2 or 3 times faster.

- A new NumCache has been setup for caching chunks during indexed
queries.  This can make repeated queries to go 2x faster, and the loss
of efficiency during non-repeated queries is very low (< 5%).



git-svn-id: http://www.pytables.org/svn/pytables/PyTablesPro/trunk@3569 1b98710c-d8ec-0310-ae81-f5f2bcd8cb94