Added a new Table.itersorted() method.  It is documented in docstrings, but not yet in UG nor the release notes.
Also, the 'allequal()' function in tables/tests/common.py is checking
now the actual dtype of arrays (before it was checking just the
'type', meaning that it was checking types for numarray, i.e. it was
outdated).  That revealed some glitches in the test suite and perhaps
in NumPy too.  I should invertigate this more.

Finally, the Index class has received new __getitem__() and
__len__() methods.  Those were necessary in order to take advantage of
the Table.itersequence() iterator for the Table.itersorted().
However, this introduced some problems because now the test:

if col.index:

is false if there are no elements in the index (!).  This has been
solved by using:

if col.is_indexed:

instead.



git-svn-id: http://www.pytables.org/svn/pytables/PyTablesPro/trunk@3726 1b98710c-d8ec-0310-ae81-f5f2bcd8cb94