Merged in r3833:3836 from trunk for sync'ing the changes needed for allowing normal behaviour when completely disable the node cache.
In this merge, there are also the needed changes to run some tests for
Pro without a node cache.  However, I'm having difficulties to run the
complete series of tests without a node cache.  In any case, a final
solution should eventually come more sooner than later (although
probably not for 2.1).

Among the additional changes in this commit, it is a new way to compute the
column of an index based only on information about the *pathname* of the
index dataset.  This is definitely a much better approach than before,
where one had to set the Index.column manually after the Index creation
time in Column constructor (very, very ugly).

All the tests seems to pass (not for a disabled cache, though :-/).



git-svn-id: http://www.pytables.org/svn/pytables/PyTablesPro/trunk@3837 1b98710c-d8ec-0310-ae81-f5f2bcd8cb94