New File.get_file_image method for file image retrieving.
The new method supersedes the old getInMemoryFileContents method that
has been removed.

The new implementation uses the H5Fget_file_image function of the HDF5
C API instead of setting custom callbacks.  The new implementation
allows a semplification of the code and fixes issues related to
updating of image files.