Add to vlarray the method get_row_size. Closes #24.
get_row_size return the total size in bytes of a given row in a
vlarray. Having this feature may allow users to know whether a read
would take a lot of memory and time or not.

Please note that this is not always indicative of how many elements
are contained in the row as there are atom of non-fixed size, e.g.
Objects. For atoms of fixed size this is equivalent to the number of
elements times the size.