Be careful with signs while interating backward.
This is a quick and possibily partial fix for #481.

It seems that C dictates that istopb - istartb - 1 is an unsigned
expression (because istartb is unsigned and istopb is signed but not big
enough to contain istartb).

With this fix I am betting that 1 - istopb + istartb is non-negative. It seems
to work but this function should be rewritten more clearly.