Fix calling File.get_node with another file.
You can call File.get_node with a Node beloging to a different file as `where`
argument. This is ok but if you also pass a `name` paramenter that name is
being looked up in the current file rather than in the file pointer by where.

This patch reworks File.get_node and fixes this behaviour.

Closes #478