Cast selection indices to integer
This is a bit weird, it was prompted by #475 and #43.

The problem (or part of it) lies in that if

    key  = numpy.uint64(10)

`is_idx(key)` will return `True` but then

    self._process_range(key, key + 1, 1)

fails because `key + 1 == 11.0` (try it yourself).

The uptake is that indices need to be casted in some way. I decided to use
`operator.index` because that's what Python uses internally when something is
used as an index.

Interestinly enough #43 got fixed accidentally: `Array._interpret_indexing`
fails because `_process_range` fails, just like above. But then
`Array.__getitem__` goes on trying different strategies and finally
`_fancy_selection` does the cast.

Closes #475, #43