Add test for encoding='latin1' to #560 testcase
py2_pickled_dict is a dictionary that contains a numpy array that has
illegal (non-ascii) characters when pickled on python 2.
It will throw a UnicodeDecodeError when unpickled on python 3, but
needs encoding='latin1' instead of encoding='bytes'. To check that
the dict is unpickled correctly we check the attribute 's' which is
u'just a string'.