Use deepcopy for recursion to work
This function returns a tuple with references to `strexpr` and
`idxexprs` which a then passed recursively to this function while
traversing an expression AST. We need to make deepcopies op these
objects to prevent overwriting them. This fixes queries such as:
((a = 1 & b = 2) | (a = 2 & b = 1)).
Fixed gh-564. Thanks to ajenkens-cargometrics for reporting.