utilsextension.encode_filename: support PEP 519
PEP 519 defines a filesystem path protocol, and enhances Python builtin I/O
functions to support "path-like" objects. As part of this PEP, the `os` module
also gained a `fspath` function, which takes an arbitrary path-like object and
returns a string.

If the os.fspath function is available, use it to convert `filename` to a
string.