Add track_times property to Leaf class
This property retrieves the "track times" creation property of a
dataset using H5Pget_obj_track_times.

Retrieval of the property works in a newly created dataset but the
property does not seem to survive closing and reopening the file
(bug?).  So, the property has dubious value.  It might be better to
just advise users to check whether the ctime for the dataset is 0.