Setup: force installation/inclusion of runtimes
In Python>=3.8 on Windows, CPython no longers searches for DLLs in
PATH. For `setup.py install` to work, we must force DLLs to be installed.
DLL are included as `data` only if found: Exit on error if DLL not found.

Wheel building (all OSes): Exit if a runtime is not found, to prevent
a wheel being build without the required runtime included.