Windows: remove pre-load of HDF5 DLLs
Preloading the HDF5 DLLs into the process via Ctypes was done to improve
diagnostics and avoid issues when loading DLLs during runtime.

This was relevant before wheels became a thing.

However, on python >= 3.8 `ctypes.cdll.LoadLibrary` does not work as expected
(does not search PATH). This means we might be able to import DLLs even when
this check fails. (This is the case for the conda built python). In the end,
we get more false positives than real failures.