CLN : Update super usage
On Python 3, super doesnt need to called with arguments where as on
Python 2, super needs to be called with a class object and an instance
as the two arguments.

Now that the pytables codebase is python 3 only, the super usage can be
updated to remove the need to pass args to super

Note that these changes were done using automated regex-based search and
replace. After the automated search and replace was done, each change
was individually checked and added to the commit. Each change was
checked to ensure that the super was being called with the current
class, self and not a super class, self.