hdf5extension: remove call to ndarray.strides
This code was added in 1da0f0bb716fa4f75b70c481fb92a4ac382afd8b.
Trying to get rid of the call to .strides is awkward, because
PyArray_DIMS() returns a plain array. I'm not sure if this is the
best way to approach this, but at least it compiles ;)