Use system includes for a few headers
This doesn't make a huge difference for the compiler because setup.py
doesn't seem to diferentiate system include paths, but it tells the
reader that this is a system header. I didn't touch the includes of
headers that are bundled (and are sometimes included from system
paths, and sometimes from internal paths).