Stop creating vl arrays with fletcher32 filter
hdf5 doesn't like that:
https://github.com/h5py/h5py/issues/1948
https://github.com/HDFGroup/hdf5/commit/16349c5fddce8a74644e18d01d7ea8186aaaa255

> The 'not suitable for filters' message does seem to have been added
> in 1.12.1 From the code, it looks like it will refuse to apply any
> filters to vlen strings.

This patch is based on the assumption that the hdf5 code is *correct*, and
indeed fletcher32 shouldn't be used in that vlarrays. But if hdf5 is wrong,
then the fix should be their side. It might make sense to apply this to
get the tests passing again, even if ultimately hdf5 is adjusted too.

Fixes https://github.com/PyTables/PyTables/issues/845.