Unify error handling in optimized array code
See comment at the beginning of the file for a summary.

While doing this, many cases of a return leaving lingering HDF5 or Blosc2
resources have been fixed, as well as cleanup blocks which closed the wrong
objects (e.g. a passed in dataset_id instead on an in-function space_id).