Remove pointless check for B2ND filter/HDF5 array type size match
Since the filter parameter is actually computed by the filter initialization
function, not read from disk, so there is no chance for an exploit.  In fact,
the filter param value is better because array types are taken into account
for compatibility with other software.

(This is also consistent with how optimized table decompression works.)