Do not pass unused output arguments to H5Dget_chunk_info
In some cases, with uniform arrays (at least with HDF5 v1.14.0), using both
"offset" and "addr" output arguments results in the latter getting an invalid
output value of 0 without an error.  Since PyTables code only uses the latter,
passing NULL as "offset" does the job.  This HDF5 behaviour is not documented,
so it may actually be a bug in there.

Applied both to table and uniform array optimized Blosc2 read for consistency.