Fix B2ND block shape computation exceeding maximum items (again)
In the case that the computed number of items is still under the maximum in
the main loop, but the inner loop hits a dimension where doubling the block
shape would exceed the chunk dimension, and raising the shape to that of the
chunk exceeds the maximum items.

A test is added to check that raising the shape does not exceed that.