Drop support for Python 3.8
At least when building and testing via CI.  This allows us to depend again on
oldest-supported-numpy at build time for increased ABI compatibility with
NumPy.  The temporary dependency on plain NumPy was caused by an
incompatibility between oldest-supported-numpy and the newer required version
of Blosc2 which only happened with Python 3.8.