Only keep current minor version notes in RELEASE_NOTES.rst
Move minor version release notes for older versions to their own file in
documentation, so that they are not lost.

The current approach was broken since the doc's release notes page had entries
with repeated titles for the latest versions, and the reader had then to
scroll down to the desired version.

Add some notes in the doc's release notes page for the last version with
instructions on what to do when a new major or minor version is released.