Fix b2nd slicing benchmark data types
It turns out that inconsistent use of dtypes in the benchmark resulted in
Blosc2 & PyTables using int64 data, while h5py used float64 data.

The types where fixed and the benchmark executed again on the same machine,
but this time with hdf5plugin 4.3.0, so that both PyTables unoptimized and
h5py use the same decompression code in the HDF5 filter pipeline, thus their
results are more similar (which makes more sense).