Fix direct chunk ops (doc & tests) to fail beyond dataset shape
The previous behaviour was not compatible with actual HDF5 operation (writing
between current and maximum shape fails) and probably based on some erroneous
tests.  At any rate, operating beyond dataset shape is unintuitive and
unreliable (at least with h5py).  The new behaviour should be easier to
understand, as any operation acting beyond dataset shape (even if it is within
maximum shape) is to fail with `ChunkError` (and never `NoSuchChunkError`,
e.g. for reads).

Implementation code has not been fixed yet.

This is a backwards-incompatible change.