Fix direct chunk ops to raise IndexError when beyond current shape
Instead of a generic `ChunkError`, since the direct use of that exception was
reduced to signaling that coordinates not in the dataset where being used.
That is the purpose of `IndexError` and thus the exception that one would
intuitively expect.

Implementation code is also fixed.

This is a backwards-incompatible change.