Fix values in missing ChunkInfo for direct chunk ops (doc & tests)
Since a `ChunkInfo` is only returned when within current dataset
shape (whether present or missing), it makes sense for it to always include
coordinates, e.g. to get the start coordinates of a missing chunk.  To ease
the detection of missing chunk info, all fields except start coordinates get a
`None` value.

Implementation code has not been fixed yet.

This is a backwards-incompatible change.